Sign in working on frontend

+++ Backend/app/Http/Controllers/DrawController.php
@@ -12,11 +12,14 @@ class DrawController extends Controller
 
     public function postDraw(Request $request){
         //confirmamos que este metodo solo se pueda ejecutar si el usuario esta logueado
+        
 
         if(! $user = JWTAuth::parseToken()->authenticate()){//authenticate() confirms that the token is valid 
             return response()->json(['message' => 'User not found'],404); //si no hay token o no es correcto lanza un error
         }
         
+        //$user = JWTAuth::parseToken()->toUser(); //Retorna el usuario del token
+
         $draw = new Draw();
         $draw->name = $request->input('name');//cogemos los datos del draw desde la request del frontend
         $draw->author = $request->input('author');
@@ -35,12 +38,14 @@ class DrawController extends Controller
         'charset' => 'utf-8'];
 
         return response()->json($response, 200, $headers);
-
-
-
     }
 
     public function putDraw(Request $request, $id){//actualizar draw atributes
+
+        if(! $user = JWTAuth::parseToken()->authenticate()){//authenticate() confirms that the token is valid 
+            return response()->json(['message' => 'User not found'],404); //si no hay token o no es correcto lanza un error
+        }
+
         $draw = Draw::find($id);
         if(!$draw){//si no ha encontrado el draw con ese id
             return response()->json(['message' => 'Draw not found'],404);//json con mensaje de error 404 not found
@@ -51,6 +56,11 @@ class DrawController extends Controller
     }
 
     public function deleteDraw($id){
+
+        if(! $user = JWTAuth::parseToken()->authenticate()){//authenticate() confirms that the token is valid 
+            return response()->json(['message' => 'User not found'],404); //si no hay token o no es correcto lanza un error
+        }
+
         $draw = Draw::find($id);
         $draw->delete();
         return response()->json(['message' => 'Draw deleted'],200);
